refactor(ecs): extract shared readEntityTile helper (DEAD-09)
bomb-tick-system.js and collision-system.js carried byte-identical copies
of readEntityTile. Move it to src/ecs/shared/tile-utils.js and import
from there in both systems. Update collision-system.test.js to import
the helper from the new location.

No behavior change. 88/88 unit tests pass.
