+++ tests/unit/systems/collision-system.test.js
@@ -20,6 +20,7 @@ import { createHealthStore } from '../../../src/ecs/components/stats.js';
 import { CELL_TYPE, GHOST_STATE } from '../../../src/ecs/resources/constants.js';
 import { createEventQueue, drain } from '../../../src/ecs/resources/event-queue.js';
 import { createMapResource, getCell } from '../../../src/ecs/resources/map-resource.js';
+import { readEntityTile } from '../../../src/ecs/shared/tile-utils.js';
 import {
   emitGameplayEvent,
   GAMEPLAY_EVENT_SOURCE,
@@ -35,7 +36,6 @@ import {
   createCollisionSystem,
   DEFAULT_GHOST_SLOTS_PER_CELL,
   isPlayerInvincible,
-  readEntityTile,
   resetCollisionScratch,
   tileToCellIndex,
 } from '../../../src/ecs/systems/collision-system.js';
